--- a/engines/badger/badger.go
+++ b/engines/badger/badger.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 
 	"github.com/rotationalio/honu/config"
+	engine "github.com/rotationalio/honu/engines"
 )
 
 func Open(conf config.ReplicaConfig) (*BadgerEngine, error) {
@@ -18,4 +19,8 @@ func (db *BadgerEngine) Engine() string {
 
 func (db *BadgerEngine) Close() error {
 	return errors.New("not implemented yet")
+}
+
+func (db *BadgerEngine) Begin(readonly bool) (engine.Transaction, error) {
+	return nil, errors.New("not implemented yet")
 }