--- a/engines/errors.go
+++ b/engines/errors.go
@@ -3,5 +3,6 @@ package engine
 import "errors"
 
 var (
-	ErrNotFound = errors.New("not found")
+	ErrNotFound   = errors.New("not found")
+	ErrReadOnlyTx = errors.New("cannot execute a write operation in a read only transaction")
 )