--- a/honu_test.go
+++ b/honu_test.go
@@ -63,7 +63,7 @@ func TestLevelDBInteractions(t *testing.T) {
 	defer db.Close()
 
 	// Put a version to the database
-	err = db.Put([]byte("foo"), []byte("this is the value of foo"))
+	_, err = db.Put([]byte("foo"), []byte("this is the value of foo"))
 	require.NoError(t, err)
 
 	// Get the version of foo from the database
@@ -77,7 +77,7 @@ func TestLevelDBInteractions(t *testing.T) {
 	require.Equal(t, uint64(1), obj.Version.Version)
 
 	// Delete the version from the database
-	err = db.Delete([]byte("foo"))
+	_, err = db.Delete([]byte("foo"))
 	require.NoError(t, err)
 
 	// Should not be able to get the deleted version
@@ -93,7 +93,7 @@ func TestLevelDBInteractions(t *testing.T) {
 	require.Empty(t, obj.Data)
 
 	// Be able to "undelete" a tombstone
-	err = db.Put([]byte("foo"), []byte("this is the undead foo"))
+	_, err = db.Put([]byte("foo"), []byte("this is the undead foo"))
 	require.NoError(t, err)
 
 	value, err = db.Get([]byte("foo"))
@@ -107,13 +107,18 @@ func TestLevelDBInteractions(t *testing.T) {
 	require.False(t, obj.Tombstone())
 
 	// Put a range of data into the database
-	require.NoError(t, db.Put([]byte("aa"), []byte("123456")))
-	require.NoError(t, db.Put([]byte("ab"), []byte("7890123")))
-	require.NoError(t, db.Put([]byte("ba"), []byte("4567890")))
-	require.NoError(t, db.Put([]byte("bb"), []byte("1234567")))
-	require.NoError(t, db.Put([]byte("bc"), []byte("9012345")))
-	require.NoError(t, db.Put([]byte("ca"), []byte("67890123")))
-	require.NoError(t, db.Put([]byte("cb"), []byte("4567890123")))
+	for _, pair := range [][]string{
+		{"aa", "123456"},
+		{"ab", "7890123"},
+		{"ba", "4567890"},
+		{"bb", "1234567"},
+		{"bc", "9012345"},
+		{"ca", "67890123"},
+		{"cb", "4567890123"},
+	} {
+		_, err = db.Put([]byte(pair[0]), []byte(pair[1]))
+		require.NoError(t, err)
+	}
 
 	// Iterate over a prefix in the database
 	iter, err := db.Iter([]byte("b"))
@@ -164,7 +169,8 @@ func BenchmarkHonuGet(b *testing.B) {
 	_, err = rand.Read(value)
 	require.NoError(b, err)
 
-	require.NoError(b, db.Put(key, value))
+	_, err = db.Put(key, value)
+	require.NoError(b, err)
 
 	// Reset the timer to focus only on the get call
 	b.ResetTimer()
@@ -227,7 +233,7 @@ func BenchmarkHonuPut(b *testing.B) {
 	// Reset the timer to focus only on the get call
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		gErr = db.Put(key, value)
+		_, gErr = db.Put(key, value)
 	}
 
 	require.NoError(b, gErr)
@@ -282,9 +288,10 @@ func BenchmarkHonuDelete(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		b.StopTimer()
-		require.NoError(b, db.Put(key, value))
+		_, err = db.Put(key, value)
+		require.NoError(b, err)
 		b.StartTimer()
-		gErr = db.Delete(key)
+		_, gErr = db.Delete(key)
 	}
 
 	require.NoError(b, gErr)
@@ -338,7 +345,8 @@ func BenchmarkHonuIter(b *testing.B) {
 		_, err = rand.Read(value)
 		require.NoError(b, err)
 
-		require.NoError(b, db.Put([]byte(key), value))
+		_, err = db.Put([]byte(key), value)
+		require.NoError(b, err)
 	}
 
 	// Reset the timer to focus only on the get call
@@ -418,7 +426,8 @@ func BenchmarkHonuObject(b *testing.B) {
 	_, err = rand.Read(value)
 	require.NoError(b, err)
 
-	require.NoError(b, db.Put(key, value))
+	_, err = db.Put(key, value)
+	require.NoError(b, err)
 
 	// Reset the timer to focus only on the get call
 	b.ResetTimer()