--- a/options/options_test.go
+++ b/options/options_test.go
@@ -9,6 +9,20 @@ import (
 	ldb "github.com/syndtr/goleveldb/leveldb/opt"
 )
 
+func TestHonuOptions(t *testing.T) {
+	// Test default options
+	opts, err := options.New()
+	require.NoError(t, err, "could not create options")
+	require.Equal(t, "default", opts.Namespace)
+	require.False(t, opts.Destroy)
+
+	// Test setting multiple options
+	opts, err = options.New(options.WithDestroy(), options.WithNamespace("foo"))
+	require.NoError(t, err, "could not create options")
+	require.Equal(t, "foo", opts.Namespace)
+	require.True(t, opts.Destroy)
+}
+
 func TestLevelDBReadOptions(t *testing.T) {
 	readOptions := &ldb.ReadOptions{
 		DontFillCache: true,
@@ -17,7 +31,7 @@ func TestLevelDBReadOptions(t *testing.T) {
 	cfg := &options.Options{}
 	ldbReadFunc := options.WithLeveldbRead(readOptions)
 	ldbReadFunc(cfg)
-	require.Equal(t, cfg.LeveldbRead, readOptions)
+	require.Equal(t, cfg.LevelDBRead, readOptions)
 }
 
 func TestLevelDBWriteOptions(t *testing.T) {
@@ -28,7 +42,7 @@ func TestLevelDBWriteOptions(t *testing.T) {
 	cfg := &options.Options{}
 	ldbWriteFunc := options.WithLeveldbWrite(writeOptions)
 	ldbWriteFunc(cfg)
-	require.Equal(t, cfg.LeveldbWrite, writeOptions)
+	require.Equal(t, cfg.LevelDBWrite, writeOptions)
 
 }