This commit fixes a pickle we were in with a production implementation that uses Honu (even though Honu is still in a prototype phase). The following changes were made:

1. Implementing a basic Transaction interface to ensure that multiple access operations in `Put` and `Delete` are consistent.
2. Implementing a `WithNamespace()` option to use more than the default namespace.
3. Returning object metadata from `Put` and `Delete`
4. Updating the leveldb implementation to use a global lock for read/write transactions
5. Updating the leveldb implementation to use `namespace::key` prefixes for namespace storage.