@tl6kk 2021-11-23
Benchmark results before PR #9 

```
BenchmarkHonuGet-8		849062		1427 ns/op
BenchmarkLevelDBGet-8		2105570		569.2 ns/op 	
BenchmarkHonuPut-8		15304		74503 ns/op 	
BenchmarkLevelDBPut-8		23002		58611 ns/op	
BenchmarkHonuDelete-8		50568		24840 ns/op	
BenchmarkLevelDBDelete-8	283647		4004 ns/op	
BenchmarkHonuIter-8		91710		11033 ns/op		
BenchmarkLevelDBIter-8		464310		2585 ns/op		
BenchmarkHonuObject-8		858931		1358 ns/op	
```

![before](https://user-images.githubusercontent.com/15382905/142983228-7da3378d-5e02-4f57-968b-bea936e33a94.png)

Benchmark results after PR #9 
```
BenchmarkHonuGet-8		816619		1454 ns/op		
BenchmarkLevelDBGet-8		2024266		564.2 ns/op	
BenchmarkHonuPut-8		15522		72811 ns/op	
BenchmarkLevelDBPut-8		23102		58457 ns/op
BenchmarkHonuDelete-8		49892		27404 ns/op	
BenchmarkLevelDBDelete-8	286095		3671 ns/op
BenchmarkHonuIter-8		89802		11757 ns/op
BenchmarkLevelDBIter-8		464150		2548 ns/op
BenchmarkHonuObject-8		804229		1435 ns/op
```
![after](https://user-images.githubusercontent.com/15382905/142983281-a82720d5-0ff6-4149-b11b-380592f0fbcb.png)

Before - After:

```
	Honu		LevelDB
Get	-27 ns/op	5 ns/op
Put	1692 ns/op	154 ns/op
Delete	-2564 ns/op	333 ns/op
Iter	-724 ns/op	37 ns/op
```